shifted query from App.tsx to Search.tsx
